Uploaded Basketball Reference Scraper
This is the code to scrape all teams' stats from a given year and then upload them to the SQL database. Unfortunately, because BR uses cloudflare protection, the code has a built-in time delay, and scraping a year takes about 25 minutes instead of being done in seconds. This time is passive, though -- just have the code running in the background while doing something else and it will work.

Apologies for code formatting -- it's a bit messy since I rushed some things -- wish it were prettier. 

To scrape a years results, all you have to do is execute scrapeYear() with the proper inputs. Make sure a table has been created with the proper columns and formatting on the sql database before scraping a year.

If you have any questions, send me a slack message.
-David